% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrationCurve.R
\name{CalibrationCurve}
\alias{CalibrationCurve}
\title{calibration model}
\usage{
CalibrationCurve(Ck, A)
}
\arguments{
\item{Ck}{Known concentrations (standards)}

\item{A}{Absorbance observed on spectrophotometer}
}
\value{
returns a linear calibration model and plot
}
\description{
This function determines fits a linear model between absorbance and known
concentrations to determine the calibration curve equation
}
