% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrationCurve.R, R/CalibrationPlot.R
\name{CalibrationPlot}
\alias{CalibrationPlot}
\title{calibration model}
\usage{
CalibrationPlot(Ck, A, cor_lab_x, cor_lab_y, eq_lab_x, eq_lab_y, file_name)

CalibrationPlot(Ck, A, cor_lab_x, cor_lab_y, eq_lab_x, eq_lab_y, file_name)
}
\arguments{
\item{Ck}{Known concentrations (standards)}

\item{A}{Absorbance observed on spectrophotometer}

\item{Ck}{Known concentrations (standards)}

\item{A}{Absorbance observed on spectrophotometer}
}
\value{
returns a linear calibration model and plot

returns a linear calibration model plot
}
\description{
This function determines fits a linear model between absorbance and known
concentrations to determine the calibration curve equation

This function determines fits a linear model between absorbance and known
concentrations to determine the calibration curve equation
}
