% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreundlichParameters.R
\name{FreundlichParameters}
\alias{FreundlichParameters}
\title{Estimate Parameters of Freundlich model}
\usage{
FreundlichParameters(Ce, Qe, output_fname)
}
\arguments{
\item{Ce}{equilibrium solution concentrations in mg/l}

\item{Qe}{retention by solid (adsorption) in mg/kg}

\item{output_fname}{file name/path to which the parameters will be exported (e.g: "analysis/langmuir.csv" )}
}
\value{
A csv file containing the estimated parameters
}
\description{
This function estimates Freundlich isotherm model parameters using the linear form of Freundlich equation.
}
