% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LangmuirParameters.R
\name{LangmuirParameters}
\alias{LangmuirParameters}
\title{Estimate Parameters of Langmuir model}
\usage{
LangmuirParameters(Ce, Qe, output_fname)
}
\arguments{
\item{Ce}{equilibrium solution concentrations in mg/l}

\item{Qe}{retention by solid (adsorption) in mg/kg}

\item{output_fname}{}
}
\value{
A csv file containing the estimated parameters
}
\description{
This function estimates langmuir isotherm model parameters using the linear form of langmuir equation.
}
