% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearPlot.R
\name{LinearPlot}
\alias{LinearPlot}
\title{Plot the Linear model fit to the experimental data}
\usage{
LinearPlot(Ce, Qe, cor_lab_x, cor_lab_y, eq_lab_x, eq_lab_y, file_name)
}
\arguments{
\item{Ce}{equilibrium solution concentrations in mg/l}

\item{Qe}{retention by solid (adsorption) mg/kg}

\item{cor_lab_x, cor_lab_y}{location on the plot to place pearson r and p-value}

\item{eq_lab_x, eq_lab_y}{location on the plot to place equation of the fitted line}

\item{file_name}{file name/path to which the plot will be exported (e.g: "Plots/Langmuir_plt" )}
}
\value{
A pdf and a png containing the graphics
}
\description{
This function plots the Linear isotherm model fit
}
