% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonLinearFreundlichPlot.R
\name{NonLinearFreundlichPlot}
\alias{NonLinearFreundlichPlot}
\title{Fits a non linear form of Langmuir model to the Ce and Qe data}
\usage{
NonLinearFreundlichPlot(df, Ce, Qe, n_guess, Kf_guess)
}
\arguments{
\item{df}{data frame containing the data}

\item{Ce}{equilibrium solution concentrations in mg/l}

\item{Qe}{retention by solid (adsorption) in mg/kg}

\item{n_guess}{Guestimate of n}

\item{Kf_guess}{Guestimate of Kf}
}
\value{
A plot of predicted non linear form of Langmuire model
with the Qmax (maximum sorption capacity) and Kl (langmuir constant
pertaining to the affinity of substrate towards the chemical) values
}
\description{
This function determines the amount of chemical (Ce) in solution
}
