% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonLinearLangmuirPlot.R
\name{NonLinearLangmuirPlot}
\alias{NonLinearLangmuirPlot}
\title{Fits a non linear form of Langmuir model to the Ce and Qe data}
\usage{
NonLinearLangmuirPlot(nlm_model, Ce, Qe)
}
\arguments{
\item{nlm_model}{Output of the nlm() on the dataset}

\item{Ce}{equilibrium solution concentrations in mg/l}

\item{Qe}{retention by solid (adsorption) in mg/kg}
}
\value{
A plot of predicted non linear form of Langmuire model
with the Qmax (maximum sorption capacity) and Kl (langmuir constant
pertaining to the affinity of substrate towards the chemical) values
}
\description{
This function determines the amount of chemical (Ce) in solution
}
