% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleAbstoConc.R
\name{SampleAbstoConc}
\alias{SampleAbstoConc}
\title{Calculate Concentrations in Samples based on Calibration curve}
\usage{
SampleAbstoConc(SampleAbs, CalibSlope, CalibIntercept, DF = 1)
}
\arguments{
\item{SampleAbs}{Absorbance values for samples}

\item{CalibSlope}{Slope of the calibration curve}

\item{CalibIntercept}{Intercept of the calibration curve}

\item{DF}{Dilution factor. Default set to 1}
}
\value{
Concentration of chemical in solution (mg/L)
}
\description{
This function determines the amount of chemical (Ce) in solution
}
