% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MassSorbed.R
\name{SorbedMass}
\alias{SorbedMass}
\title{Calculate amount of chemical adsorbed/retained by solid (q)}
\usage{
SorbedMass(Ci, Cf, V = 0.025, m = 0.001)
}
\arguments{
\item{Ci}{Initial concentration of the solution (ppm)}

\item{Cf}{Final concentration of the solution (ppm)}

\item{V}{Volumne of solution in Liters}

\item{m}{Mass of solid in Kilograms}
}
\value{
Mass of chemical sorbed to solid (mg/kg)
}
\description{
This function determines the amount of chemical (q) in solution adsorbed
by the solid
}
