% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrationPlot.R
\name{CalibrationPlot}
\alias{CalibrationPlot}
\title{Calibration model fit plot}
\usage{
CalibrationPlot(Ck, A, cor_lab_x, cor_lab_y, eq_lab_x, eq_lab_y)
}
\arguments{
\item{Ck}{Known concentrations (standards)}

\item{A}{Absorbance observed on spectrophotometer}
}
\value{
returns a linear calibration model plot
}
\description{
This function determines a linear model between absorbance and known
concentrations to determine the calibration curve equation
}
